# docs/open_api_spec.json
{"openapi":"3.0.1","info":{"title":"Card Auth Server ACS","description":"Card Auth Server ACS","contact":{"name":"Freedom Finance Stack","email":"contact@freedomfinancestack.org"},"license":{"name":"Apache 2.0","url":"https://www.apache.org/licenses/LICENSE-2.0.html"},"version":"1.0.0"},"servers":[{"url":"http://127.0.0.1:8080","description":"Generated server url"}],"paths":{"/v2/transaction/challenge/browser":{"post":{"tags":["challenge-request-controller"],"summary":"Handles browser Challenge Request generating user's browser","operationId":"handleChallengeRequest","parameters":[{"name":"creq","in":"query","required":true,"schema":{"type":"string"}},{"name":"threeDSSessionData","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"Request Successfully handled and validated","content":{"html/text;charset=utf-8":{"schema":{"type":"string"}}}},"400":{"description":"Bad Request or Request not according to Areq Schema","content":{"html/text;charset=utf-8":{"schema":{"type":"string"}}}},"500":{"description":"Server Exception Occurred during request handling","content":{"html/text;charset=utf-8":{"schema":{"type":"string"}}}}}}},"/v2/transaction/challenge/browser/validate":{"post":{"tags":["challenge-request-controller"],"summary":"Handles browser validation Challenge Request generating user's browser","operationId":"handleChallengeValidationRequest","requestBody":{"content":{"application/x-www-form-urlencoded;charset=UTF-8":{"schema":{"$ref":"#/components/schemas/CVReq"}}}},"responses":{"200":{"description":"Request Successfully handled and validated","content":{"html/text;charset=utf-8":{"schema":{"type":"string"}}}},"400":{"description":"Bad Request or Request not according to Areq Schema","content":{"html/text;charset=utf-8":{"schema":{"type":"string"}}}},"500":{"description":"Server Exception Occurred during request handling","content":{"html/text;charset=utf-8":{"schema":{"type":"string"}}}}}}},"/v2/transaction/challenge/app":{"post":{"tags":["challenge-request-controller"],"summary":"Handles App Based Challenge Request","operationId":"handleChallengeRequest_1","requestBody":{"content":{"application/jose;charset=utf-8":{"schema":{"$ref":"#/components/schemas/CREQ"}},"application/jose":{"schema":{"$ref":"#/components/schemas/CREQ"}},"application/json;charset=utf-8":{"schema":{"$ref":"#/components/schemas/CREQ"}}},"required":true},"responses":{"200":{"description":"Request Successfully handled and validated","content":{"application/jose;charset=UTF-8":{"schema":{"$ref":"#/components/schemas/CRES"}}}},"400":{"description":"Bad Request or Request not according to CReq Schema","content":{"application/jose;charset=UTF-8":{"schema":{"$ref":"#/components/schemas/CRES"}}}},"500":{"description":"Server Exception Occurred during request handling","content":{"application/jose;charset=UTF-8":{"schema":{"$ref":"#/components/schemas/CRES"}}}}}}},"/v2/transaction/authentication":{"post":{"tags":["authentication-request-controller"],"summary":"Handles Authentication Request generating from 3DS Server","operationId":"handleAuthenticationRequest","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AREQ"}}},"required":true},"responses":{"200":{"description":"Request Successfully handled and validated","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ARES"}}}},"400":{"description":"Bad Request or Request not according to Areq Schema","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ARES"}}}},"500":{"description":"Server Exception Occurred during request handling","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ARES"}}}}}}}},"components":{"schemas":{"CVReq":{"type":"object","properties":{"transactionId":{"type":"string"},"authVal":{"type":"string"},"resendChallenge":{"type":"boolean"},"cancelChallenge":{"type":"boolean"},"threeDSMessageType":{"type":"string","enum":["AReq","ARes","CReq","CRes","Erro","RReq","CDRes","CVReq","RRes"]}}},"CREQ":{"type":"object","properties":{"threeDSServerTransID":{"type":"string"},"threeDSRequestorAppURL":{"type":"string"},"acsTransID":{"type":"string"},"challengeWindowSize":{"type":"string"},"messageType":{"type":"string"},"messageVersion":{"type":"string"},"sdkCounterStoA":{"type":"string"},"sdkTransID":{"type":"string"},"challengeCancel":{"type":"string"},"challengeDataEntry":{"type":"string"},"challengeHTMLDataEntry":{"type":"string"},"messageExtension":{"type":"array","items":{"$ref":"#/components/schemas/MessageExtension"}},"resendChallenge":{"type":"string"},"challengeNoEntry":{"type":"string"},"threeDSMessageType":{"type":"string","enum":["AReq","ARes","CReq","CRes","Erro","RReq","CDRes","CVReq","RRes"]}}},"MessageExtension":{"type":"object","properties":{"name":{"type":"string"},"id":{"type":"string"},"criticalityIndicator":{"type":"boolean"},"data":{"type":"object","additionalProperties":{"type":"object"}},"valid":{"type":"boolean"}}},"CRES":{"type":"object","properties":{"threeDSServerTransID":{"type":"string"},"acsCounterAtoS":{"type":"string"},"acsTransID":{"type":"string"},"acsUiType":{"type":"string"},"challengeCompletionInd":{"type":"string"},"messageType":{"type":"string"},"messageVersion":{"type":"string"},"sdkTransID":{"type":"string"},"acsHTMLRefresh":{"type":"string"},"challengeAddInfo":{"type":"string"},"challengeInfoHeader":{"type":"string"},"challengeInfoLabel":{"type":"string"},"challengeInfoText":{"type":"string"},"challengeInfoTextIndicator":{"type":"string"},"challengeSelectInfo":{"type":"array","items":{"$ref":"#/components/schemas/ChallengeSelectInfo"}},"expandInfoLabel":{"type":"string"},"expandInfoText":{"type":"string"},"oobAppURL":{"type":"string"},"acsHTML":{"type":"string"},"issuerImage":{"$ref":"#/components/schemas/Image"},"messageExtension":{"type":"array","items":{"$ref":"#/components/schemas/MessageExtension"}},"oobAppLabel":{"type":"string"},"oobContinue":{"type":"string"},"oobContinueLabel":{"type":"string"},"psImage":{"$ref":"#/components/schemas/Image"},"resendInformationLabel":{"type":"string"},"submitAuthenticationLabel":{"type":"string"},"transStatus":{"type":"string"},"whyInfoLabel":{"type":"string"},"whyInfoText":{"type":"string"},"whitelistingInfoText":{"type":"string"},"threeDSMessageType":{"type":"string","enum":["AReq","ARes","CReq","CRes","Erro","RReq","CDRes","CVReq","RRes"]}}},"ChallengeSelectInfo":{"type":"object","properties":{"phone":{"type":"string"},"mail":{"type":"string"},"sms":{"type":"string"},"email":{"type":"string"}}},"Image":{"type":"object","properties":{"medium":{"type":"string"},"high":{"type":"string"},"extraHigh":{"type":"string"}}},"AREQ":{"type":"object","properties":{"threeDSRequestorURL":{"type":"string"},"threeDSCompInd":{"type":"string"},"threeDSRequestorAuthenticationInd":{"type":"string"},"threeDSRequestorID":{"type":"string"},"threeDSRequestorName":{"type":"string"},"threeDSServerRefNumber":{"type":"string"},"threeDSServerTransID":{"type":"string"},"threeDSServerURL":{"type":"string"},"threeRIInd":{"type":"string"},"acquirerBIN":{"type":"string"},"acquirerMerchantID":{"type":"string"},"browserAcceptHeader":{"type":"string"},"browserJavaEnabled":{"type":"string"},"browserLanguage":{"type":"string"},"browserColorDepth":{"type":"string"},"browserScreenHeight":{"type":"string"},"browserScreenWidth":{"type":"string"},"browserTZ":{"type":"string"},"browserUserAgent":{"type":"string"},"acctNumber":{"type":"string"},"deviceChannel":{"type":"string"},"deviceRenderOptions":{"$ref":"#/components/schemas/DeviceRenderOptions"},"mcc":{"type":"string"},"merchantCountryCode":{"type":"string"},"merchantName":{"type":"string"},"messageCategory":{"type":"string"},"messageType":{"type":"string"},"messageVersion":{"type":"string"},"notificationURL":{"type":"string"},"purchaseAmount":{"type":"string"},"purchaseCurrency":{"type":"string"},"purchaseExponent":{"type":"string"},"purchaseDate":{"type":"string"},"sdkAppID":{"type":"string"},"sdkEphemPubKey":{"$ref":"#/components/schemas/EphemPubKey"},"sdkMaxTimeout":{"type":"string"},"sdkReferenceNumber":{"type":"string"},"sdkTransID":{"type":"string"},"threeDSRequestorAuthenticationInfo":{"$ref":"#/components/schemas/ThreeDSRequestorAuthenticationInfo"},"threeDSRequestorChallengeInd":{"type":"string"},"threeDSRequestorPriorAuthenticationInfo":{"$ref":"#/components/schemas/ThreeDSRequestorPriorAuthenticationInfo"},"addrMatch":{"type":"string"},"acctInfo":{"$ref":"#/components/schemas/CardholderAccountInformation"},"acctID":{"type":"string"},"merchantRiskIndicator":{"$ref":"#/components/schemas/ThreeDSMerchantFeilds"},"threeDSServerOperatorID":{"type":"string"},"acctType":{"type":"string"},"broadInfo":{"$ref":"#/components/schemas/BrodInfo"},"browserIP":{"type":"string"},"cardExpiryDate":{"type":"string"},"billAddrCity":{"type":"string"},"billAddrCountry":{"type":"string"},"billAddrLine1":{"type":"string"},"billAddrLine2":{"type":"string"},"billAddrLine3":{"type":"string"},"billAddrPostCode":{"type":"string"},"billAddrState":{"type":"string"},"email":{"type":"string"},"homePhone":{"$ref":"#/components/schemas/Phone"},"mobilePhone":{"$ref":"#/components/schemas/Phone"},"cardholderName":{"type":"string"},"shipAddrCity":{"type":"string"},"shipAddrCountry":{"type":"string"},"shipAddrLine1":{"type":"string"},"shipAddrLine2":{"type":"string"},"shipAddrLine3":{"type":"string"},"shipAddrPostCode":{"type":"string"},"shipAddrState":{"type":"string"},"workPhone":{"$ref":"#/components/schemas/Phone"},"deviceInfo":{"type":"string"},"dsReferenceNumber":{"type":"string"},"dsTransID":{"type":"string"},"dsURL":{"type":"string"},"payTokenInd":{"type":"string"},"purchaseInstalData":{"type":"string"},"messageExtension":{"type":"array","items":{"$ref":"#/components/schemas/MessageExtension"}},"recurringExpiry":{"type":"string"},"recurringFrequency":{"type":"string"},"sdkEncData":{"type":"string"},"transType":{"type":"string"},"threeDSReqAuthMethodInd":{"type":"string"},"threeDSRequestorDecMaxTime":{"type":"string"},"threeDSRequestorDecReqInd":{"type":"string"},"browserJavascriptEnabled":{"type":"string"},"payTokenSource":{"type":"string"},"whiteListStatus":{"type":"string"},"whiteListStatusSource":{"type":"string"},"threeDSMessageType":{"type":"string","enum":["AReq","ARes","CReq","CRes","Erro","RReq","CDRes","CVReq","RRes"]},"transactionId":{"type":"string"}}},"BrodInfo":{"type":"object","properties":{"data":{"type":"string"}}},"CardholderAccountInformation":{"type":"object","properties":{"chAccAgeInd":{"type":"string"},"chAccDate":{"type":"string"},"chAccChangeInd":{"type":"string"},"chAccChange":{"type":"string"},"chAccPwChangeInd":{"type":"string"},"chAccPwChange":{"type":"string"},"shipAddressUsageInd":{"type":"string"},"shipAddressUsage":{"type":"string"},"txnActivityDay":{"type":"string"},"txnActivityYear":{"type":"string"},"provisionAttemptsDay":{"type":"string"},"nbPurchaseAccount":{"type":"string"},"suspiciousAccActivity":{"type":"string"},"shipNameIndicator":{"type":"string"},"paymentAccInd":{"type":"string"},"paymentAccAge":{"type":"string"},"length":{"type":"integer","format":"int32"},"empty":{"type":"boolean"},"valid":{"type":"boolean"},"dataValid":{"type":"boolean"}}},"DeviceRenderOptions":{"type":"object","properties":{"sdkInterface":{"type":"string"},"sdkUiType":{"type":"array","items":{"type":"string"}},"valid":{"type":"boolean"},"mandatoryValueAvailable":{"type":"boolean"}}},"EphemPubKey":{"type":"object","properties":{"alg":{"type":"string"},"kid":{"type":"string"},"use":{"type":"string"},"kty":{"type":"string"},"crv":{"type":"string"},"x":{"type":"string"},"y":{"type":"string"},"valid":{"type":"boolean"}}},"Phone":{"type":"object","properties":{"cc":{"type":"string"},"subscriber":{"type":"string"},"valid":{"type":"boolean"}}},"ThreeDSMerchantFeilds":{"type":"object","properties":{"shipIndicator":{"type":"string"},"deliveryTimeframe":{"type":"string"},"deliveryEmailAddress":{"type":"string"},"reorderItemsInd":{"type":"string"},"preOrderPurchaseInd":{"type":"string"},"preOrderDate":{"type":"string"},"giftCardAmount":{"type":"string"},"giftCardCurr":{"type":"string"},"giftCardCount":{"type":"string"},"length":{"type":"integer","format":"int32"},"valid":{"type":"boolean"}}},"ThreeDSRequestorAuthenticationInfo":{"type":"object","properties":{"threeDSReqAuthMethod":{"type":"string"},"threeDSReqAuthTimestamp":{"type":"string"},"threeDSReqAuthData":{"type":"string"},"valid":{"type":"boolean"}}},"ThreeDSRequestorPriorAuthenticationInfo":{"type":"object","properties":{"threeDSReqPriorAuthData":{"type":"string"},"threeDSReqPriorAuthMethod":{"type":"string"},"threeDSReqPriorAuthTimestamp":{"type":"string"},"threeDSReqPriorRef":{"type":"string"},"valid":{"type":"boolean"},"mandatoryValueAvailable":{"type":"boolean"}}},"ACSRenderingType":{"type":"object","properties":{"acsInterface":{"type":"string"},"acsUiTemplate":{"type":"string"}}},"ARES":{"type":"object","properties":{"threeDSServerTransID":{"type":"string"},"acsReferenceNumber":{"type":"string"},"acsTransID":{"type":"string"},"dsReferenceNumber":{"type":"string"},"dsTransID":{"type":"string"},"messageType":{"type":"string"},"messageVersion":{"type":"string"},"sdkEphemPubKey":{"type":"string"},"sdkTransID":{"type":"string"},"transStatus":{"type":"string"},"cardholderInfo":{"type":"string"},"acsChallengeMandated":{"type":"string"},"acsOperatorID":{"type":"string"},"acsRenderingType":{"$ref":"#/components/schemas/ACSRenderingType"},"acsSignedContent":{"type":"string"},"acsURL":{"type":"string"},"authenticationType":{"type":"string"},"authenticationValue":{"type":"string"},"broadInfo":{"type":"string"},"eci":{"type":"string"},"messageExtension":{"type":"array","items":{"$ref":"#/components/schemas/MessageExtension"}},"transStatusReason":{"type":"string"},"acsDecConInd":{"type":"string"},"whiteListStatus":{"type":"string"},"whiteListStatusSource":{"type":"string"},"threeDSMessageType":{"type":"string","enum":["AReq","ARes","CReq","CRes","Erro","RReq","CDRes","CVReq","RRes"]}}}},"securitySchemes":{"bearerAuth":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}}}}